requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
distro==1.9.0
ecdsa==0.19.0
Flask==2.0.1
Flask-Cors==4.0.0
groq==0.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.3.4
MarkupSafe==3.0.2
pyasn1==0.6.1
pydantic==2.10.1
pydantic_core==2.27.1
python-dotenv==0.19.0
python-jose==3.3.0
python-slugify==5.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
text-unidecode==1.3
typing_extensions==4.12.2
Werkzeug==2.2.2